backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
anthropic
pymupdf
pydantic
pydantic-settings
python-dotenv
